fix failures of Servicetests after adding security layer by adding authentication and create UI layer as well as nosql

+++ src/main/java/com/group6/project/relational/account/AccountCurrencyRepository.java
@@ -6,6 +6,7 @@ import java.util.List;
 
 public interface AccountCurrencyRepository extends JpaRepository<AccountCurrency, Long>{
 
-    public List<AccountCurrency> findByAccountID(Integer id);
+    public List<AccountCurrency> findByAccountID(Long id);
+
 
 }
